Refactoring and moving it towards a library
